% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendsegmentR-package.R
\docType{package}
\name{trendsegmentR-package}
\alias{trendsegmentR}
\alias{trendsegmentR-package}
\title{trendsegmentR: Linear Trend Segmentation}
\description{
Performs the detection of linear trend changes for univariate time series by implementing the bottom-up unbalanced wavelet transformation proposed by H. Maeng and P. Fryzlewicz (2023). The estimated number and locations of the change-points are returned with the piecewise-linear estimator for signal.
}
\author{
\strong{Maintainer}: Hyeyoung Maeng \email{hyeyoung.maeng@durham.ac.uk}

Authors:
\itemize{
  \item Piotr Fryzlewicz \email{p.fryzlewicz@lse.ac.uk}
}

}
\keyword{internal}
