% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendsegmentR.R
\docType{package}
\name{trendsegmentR}
\alias{trendsegmentR}
\alias{trendsegmentR-package}
\title{trendsegmentR: A device for detecting multiple change-points corresponding to linear trend changes or point anomalies in one dimensional data}
\description{
Performs the detection of point anomalies and linear trend changes by transforming the data through an adaptively constructed unbalanced wavelet basis in a hierarchical way. To start with, see the function \code{trendsegment}.
}
\references{
H. Maeng and P. Fryzlewicz (2019), Detecting linear trend changes and point anomalies in data sequences, preprint (\url{http://personal.lse.ac.uk/maengh/}).
}
\seealso{
\code{\link{trendsegment}}, \code{\link{TGUW}}, \code{\link{thresholding}}, \code{\link{invTGUW}}
}
\author{
Hyeyoung Maeng, \email{h.maeng@lse.ac.uk}
}
