% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendsegmentR.R
\docType{package}
\name{trendsegmentR}
\alias{trendsegmentR}
\title{trendsegmentR: A device for detecting multiple change-points corresponding to linear trend changes in one dimensional data}
\description{
Performs the detection of linear trend changes by transforming the data through an adaptively constructed unbalanced wavelet basis in a hierarchical way. To start with, see the function \code{trendsegment}.
}
\references{
H. Maeng and P. Fryzlewicz (2023), Detecting linear trend changes in data sequences.
}
\seealso{
\code{\link{trendsegment}}, \code{\link{TGUW}}, \code{\link{thresholding}}, \code{\link{invTGUW}}
}
\author{
Hyeyoung Maeng \email{hyeyoung.maeng@durham.ac.uk}, Piotr Fryzlewicz \email{p.fryzlewicz@lse.ac.uk}
}
